# Optimize ReadDir iterator